Reorganize data folders
This also renames the data source files
