+++ app/src/main/java/com/example/ihsastable/data/datasource/UserRemoteTestDataSource.java
@@ -1,0 +1,4 @@
+package com.example.ihsastable.data.datasource;
+
+public class UserRemoteTestDataSource {
+}
