+++ app/src/main/java/com/example/ihsastable/data/model/Coach.java
@@ -1,0 +1,4 @@
+package com.example.ihsastable.data.model;
+
+public class Coach {
+}
