+++ app/src/main/java/com/example/ihsastable/data/model/User.java
@@ -1,0 +1,4 @@
+package com.example.ihsastable.data.model;
+
+public class User {
+}
