+++ app/src/main/java/com/example/ihsastable/data/repository/EventRepository.java
@@ -1,0 +1,4 @@
+package com.example.ihsastable.data.repository;
+
+public class EventRepository {
+}
