+++ app/src/main/java/com/example/ihsastable/data/repository/RiderRepository.java
@@ -1,0 +1,4 @@
+package com.example.ihsastable.data.repository;
+
+public class RiderRepository {
+}
