+++ app/src/main/java/com/example/ihsastable/data/repository/UserRepository.java
@@ -1,0 +1,4 @@
+package com.example.ihsastable.data.repository;
+
+public class UserRepository {
+}
