--- app/src/main/java/com/example/ihsastable/datasource/UserRemoteDataSource.java
@@ -1,4 +1,0 @@
-package com.example.ihsastable.datasource;
-
-public class UserRemoteDataSource {
-}
