--- app/src/main/java/com/example/ihsastable/model/Event.java
@@ -1,4 +1,0 @@
-package com.example.ihsastable.model;
-
-public class Event {
-}
