--- app/src/main/java/com/example/ihsastable/repository/EventAdminRepository.java
@@ -1,4 +1,0 @@
-package com.example.ihsastable.repository;
-
-public class EventAdminRepository {
-}
