--- app/src/main/java/com/example/ihsastable/repository/EventRepository.java
@@ -1,4 +1,0 @@
-package com.example.ihsastable.repository;
-
-public class EventRepository {
-}
