--- app/src/main/java/com/example/ihsastable/repository/HorsesRepository.java
@@ -1,4 +1,0 @@
-package com.example.ihsastable.repository;
-
-public class HorsesRepository {
-}
